[wip][fix] fix type of ppn (u32 -> usize)

+++ src/cpu/mmu.rs
@@ -26,7 +26,7 @@ impl MMU {
         const PAGESIZE: usize = 4096; // 2^12
 
         let satp = self.csrs.borrow().read(CSRname::satp.wrap());
-        let ppn = satp & 0xFFFFF3;
+        let ppn = (satp & 0xFFFFF3) as usize;
         let state = match satp >> 31 & 0x1 {
             1 => AddrTransMode::Sv32,
             _ => AddrTransMode::Bare,
